requirements.txt
loguru==0.6.0
pyTelegramBotAPI==4.4.0
python-dotenv==0.20.0
requests==2.27.1
pandas==1.4.1
peewee==3.14.10
